+++ src/semanticembed/__init__.py
@@ -17,7 +17,7 @@ from .dedupe import dedupe_edges
 from .explain import explain, ask
 from .find_edges import find_edges
 
-__version__ = "0.6.0"
+__version__ = "0.7.0"
 
 # Set this to your license key to unlock unlimited nodes:
 #   import semanticembed